Refactor: make PrintOutputOnFailureWatcher available in other tests
